User commit

+++ src/main/java/com/example/demo/controller/ValidationController.java
@@ -19,7 +19,7 @@ public class ValidationController{
     }
 
     @GetMapping("/put/{id}")
-    public ValidationEntity getData(@PathVariable int id){
+    public ValidationEntity getData(@PathVariable Long id){
         return ser.getData(id);
     }
 }
