+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -5,5 +5,5 @@ import com.example.demo.entity.ValidationEntity;
 
 public interface ValidationService{
         ValidationEntity postdata(ValidationEntity stu);
-        ValidationEntity getData(int id);
+        ValidationEntity getData(Long id);
 }
